Remove max width from .intro
So max width was totally harshing the vibe bro:
https://drive.google.com/open?id=0B-gRZdZ8Cf_mam5tdkNmbl9ST1U
I made removed maxed width from the intro because I like the full screen no bleed type look for this, but another option that looks tight would be to put this up in the #root div:

margin: 0 auto;
max-width: 1334px;
background-color: black;

Add margin-bottom to CTA button

https://drive.google.com/open?id=0B-gRZdZ8Cf_mak5mck1OMmo4LVk

The button was chillin up on the bottom of the viewport. I ain't for that.

Suggested some other style changes and less generic call to action. Might need some more brainstorming, but you get the idea.

Bonus points if we could somehow get the image to lead one's eyes to the CTA button.